splitlunch: Fix sign of financial impact from payments
Payments and expenses should have the same signs. In a non-self-paid
expense, you owe the other person "more" money afterwards. The same is
true of a non-self-paid payment (though in that case, "more" is often
just "less negative").

This lets us unify the calculation to be agnostic to whether a
transaction is an expense or a payment.