✨ SplitLunch Refresh: Allow Self Paid + Allow Payments
* splitlunch: Fix transactions where user is uninvolved

This inverts the logic for `self_paid` in both `_get_expense_impact` and
`_get_payment_impact`, defaulting it to False. Previously, if the user
was completely uninvolved in a transaction (e.g. it was solely between
other members of a group), `self_paid` would be incorrectly set to True.
Additionally, filter these transactions out of the refresh sync.

And finally, in order to make the logic easier to test, the functions
are pulled to the top-level of the `lunchmoney_splitwise` module, so
that they can be tested without requiring the full machinery of the
SplitLunch class.

* splitlunch: Fix sign of financial impact from payments

Payments and expenses should have the same signs. In a non-self-paid
expense, you owe the other person "more" money afterwards. The same is
true of a non-self-paid payment (though in that case, "more" is often
just "less negative").

This lets us unify the calculation to be agnostic to whether a
transaction is an expense or a payment.

* splitlunch: Rework to not require personal_share

The personal_share field worked for the subset of transactions that are
currently eligible for the "splitlunch refresh" sync function, but it
would not be correct if we expanded refresh to also allow syncing
self-paid and payment transactions. Meanwhile, the only other usage of
that field was for the split in SplitLunchImport transactions, which
were always self-paid non-payments. It's easy to switch SplitLunchImport
to use financial_impact directly, and in the case of refresh sync, it's
correct for all transactions (since financial_impact represents the
change to the user's balance in Splitwise).

For simplicity, this also flips the sign of the financial_impact field.
A positive financial_impact now indicates that the user owes money to
other Splitwise users as a result of the transaction. (The old logic
effectively did this in the cases it cared about by passing most values
to `abs`.)

* splitlunch: Add --allow-{self-paid,payments} flags to refresh

These both default to disallowing these types of transactions,
preserving the out-of-box behavior from before.

* splitlunch: Add documentation about new flags

* ✨ SplitLunch Refresh: Allow Self Paid + Allow Payments

---------

Co-authored-by: Justin Flannery <juftin@juftin.com>